+++ crypto_sign/dilithium3/avx2/fips202x4.h
@@ -2,6 +2,7 @@
 #define PQCLEAN_DILITHIUM3_AVX2_FIPS202X4_H
 
 #include <immintrin.h>
+#include <stddef.h>
 #include <stdint.h>
 
 #include "params.h"
@@ -12,14 +13,14 @@ void PQCLEAN_DILITHIUM3_AVX2_shake128_absorb4x(
     const uint8_t *m1,
     const uint8_t *m2,
     const uint8_t *m3,
-    unsigned long long mlen);
+    size_t mlen);
 
 void PQCLEAN_DILITHIUM3_AVX2_shake128_squeezeblocks4x(
     uint8_t *h0,
     uint8_t *h1,
     uint8_t *h2,
     uint8_t *h3,
-    unsigned long nblocks,
+    size_t nblocks,
     __m256i *s);
 
 void PQCLEAN_DILITHIUM3_AVX2_shake256_absorb4x(
@@ -28,14 +29,14 @@ void PQCLEAN_DILITHIUM3_AVX2_shake256_absorb4x(
     const uint8_t *m1,
     const uint8_t *m2,
     const uint8_t *m3,
-    unsigned long long mlen);
+    size_t mlen);
 
 void PQCLEAN_DILITHIUM3_AVX2_shake256_squeezeblocks4x(
     uint8_t *h0,
     uint8_t *h1,
     uint8_t *h2,
     uint8_t *h3,
-    unsigned long nblocks,
+    size_t nblocks,
     __m256i *s);
 
 void PQCLEAN_DILITHIUM3_AVX2_shake128_4x(
@@ -43,23 +44,23 @@ void PQCLEAN_DILITHIUM3_AVX2_shake128_4x(
     uint8_t *h1,
     uint8_t *h2,
     uint8_t *h3,
-    unsigned long long hlen,
+    size_t hlen,
     const uint8_t *m0,
     const uint8_t *m1,
     const uint8_t *m2,
     const uint8_t *m3,
-    unsigned long long mlen);
+    size_t mlen);
 
 void PQCLEAN_DILITHIUM3_AVX2_shake256_4x(
     uint8_t *h0,
     uint8_t *h1,
     uint8_t *h2,
     uint8_t *h3,
-    unsigned long long hlen,
+    size_t hlen,
     const uint8_t *m0,
     const uint8_t *m1,
     const uint8_t *m2,
     const uint8_t *m3,
-    unsigned long long mlen);
+    size_t mlen);
 
 #endif
